apps/voice_input/voice_input.requirements.txt
# Voice Input Service Dependencies - Isolated Version

# Core FastAPI with exact versions
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.2.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Redis support
redis==5.0.1

# TTS engines
gtts==2.5.1
edge-tts==6.1.6

# Speech-to-text
openai-whisper==20231117
transformers==4.40.2
sentencepiece==0.1.99

# Intent recognition and NLP
spacy==3.7.2

# Audio enhancement
librosa==0.10.1
pydub==0.25.1

# Vision and AI Model Integration
openai==1.3.7
groq==0.4.2
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Image processing and analysis
imageio==2.31.5
scikit-image==0.21.0

# HTTP client for API calls
httpx==0.25.2
aiohttp==3.9.1

# Audio processing for TTS
soundfile==0.12.1

# Basic utilities with exact versions
python-dotenv==1.0.0
structlog==23.2.0 